+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- pager: Syntax highlighting now matches the TUI Files tab.
+
 ## [0.10.0] - 2026-07-01
 
 ### Added